Organize components into subfolders and externalize validation config
