+++ Work/src/components/headComponent/index.js
@@ -1,22 +1,5 @@
-import { validateInput } from './validation/validateInput';
-
-const HEAD_CHECKS = [
-  {
-    field: 'title',
-    errorMessage: 'no title was passed',
-    rules: ['required', 'nonEmptyString'],
-  },
-  {
-    field: 'headStyles',
-    errorMessage: 'no headStyles was passed',
-    rules: ['required', 'nonEmptyString'],
-  },
-  {
-    field: 'fonts',
-    errorMessage: 'no fonts was passed',
-    rules: ['required', 'nonEmptyString'],
-  },
-];
+import { validateInput } from '../validation/validateInput';
+import { HEAD_CHECKS } from './headComponent.config';
 
 const headComponent = (params = {}) => {
   validateInput(params, HEAD_CHECKS);
