+++ Work/src/components/index.js
@@ -1,8 +1,7 @@
 import EmailTemplateBodyComponent from './body';
 import innerContentComponent from './innerContentComponent';
-import previewTextComponent from '../components/previewText';
-
-import MainComponent from '../components/mainComponent';
+import previewTextComponent from './previewText';
+import MainComponent from './mainComponent';
 
 import { HeadHTMLString as headString } from '../display/sections/head';
