+++ Work/src/components/innerContentComponent/index.js
