+++ Work/src/components/mainComponent/mainComponent.config.js
@@ -1,0 +1,12 @@
+export const MAIN_VALIDATION_CHECKS = [
+  {
+    field: 'head',
+    errorMessage: 'no head was passed',
+    rules: ['required', 'string', 'nonEmptyString'],
+  },
+  {
+    field: 'body',
+    errorMessage: 'no body was passed',
+    rules: ['required', 'string', 'nonEmptyString'],
+  },
+];
