+++ Work/tests/unit/footer.unit.test.js
@@ -1,4 +1,4 @@
-import footerComponent from '../../src/components/footer';
+import footerComponent from '../../src/components/footer/index';
 
 const validParams = {
   address: '<span>123 Main St</span>',
