+++ Work/tests/unit/headStyles.unit.test.js
@@ -1,4 +1,4 @@
-import headStylesComponent from '../../src/components/headStyles';
+import headStylesComponent from '../../src/components/headStyles/index';
 
 describe('headStylesComponent - output validation', () => {
   test('returns a string', () => {
