+++ Work/tests/unit/previewText.unit.test.js
@@ -1,4 +1,4 @@
-import previewTextComponent from '../../src/components/previewText';
+import previewTextComponent from '../../src/components/previewText/index';
 
 describe('previewTextComponent - throws on missing or invalid content', () => {
   test('throws when content is not provided', () => {
